// src/app/feature-modules/main/sidebarconfigure/channels/data/constant-value.enum.ts
export enum Constant {
    CONTROLLER = "/SymChannel",
    PATH = "/main/configure/channels",
    UNCLICKABLE = "unclickable",
    CLICKABLE = "clickable",
    CREATE = "/create",
    EDIT = "/edit",
    DELETE = "/delete",
    SORT_BY = "channel_id",
    SORT_ORDER = "asc",
    FILTER_BY = "id",
    SAVE = "Save",
    UPDATE = "Update"
}